requirements.txt
ConfigArgParse==1.3
geomloss==0.2.4
jax==0.2.12
jaxlib==0.1.65
matplotlib==3.4.2
neural-tangents==0.3.6
numpy==1.20.1
Pillow==8.1.0
PyYAML==5.4.1
tqdm==4.56.2
torch==1.8.1
torchvision==0.9.1
